Extract export formats into a dedicated module
